backend/requirements.txt
fastapi>=0.110.0
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
motor>=3.3.0
paho-mqtt>=2.0.0
python-dotenv>=1.0.1
PyJWT[crypto]>=2.8.0
bcrypt>=4.1.2
uvicorn[standard]>=0.29.0
gunicorn>=21.2.0
python-multipart>=0.0.9
# ML integration
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
joblib>=1.3.0
scipy>=1.12.0

